requirements.txt
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
config==0.4.0
Django==2.1.3
django-storages==1.7.1
et-xmlfile==1.0.1
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-language==1.1.0
google-cloud-speech==0.36.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
httplib2==0.12.0
idna==2.7
jdcal==1.4
nltk==3.4
numpy==1.15.4
oauth2client==4.1.3
openpyxl==2.5.11
pandas==0.23.4
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
rsa==4.0
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.24.1
